Revert "refactor: 엔드포이트 URL 변경 - 추후 버저닝을 고려하여 RequestMapping에서 v1/~ 이하 부분 제거"
This reverts commit 2bd06f99389e5f9416dc167a65409924e7bf4145.
