+++ src/main/java/com/bugflix/weblog/page/PageController.java
@@ -1,19 +1,18 @@
-package com.bugflix.weblog.page.controller;
+package com.bugflix.weblog.page;
 
-import com.bugflix.weblog.page.domain.Page;
-import com.bugflix.weblog.page.service.PageServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/v1/page")
 @RequiredArgsConstructor
 @Slf4j
 public class PageController {
     private final PageServiceImpl pageService;
 
-    @GetMapping("/v1/page")
+    @GetMapping
     public Page getPage(@RequestParam(name = "url") String url) throws Exception {
 
         return pageService.getPage(url);
